+++ api/go.mod
@@ -5,7 +5,6 @@ go 1.16
 require (
 	github.com/cosmos/cosmos-sdk v0.44.3
 	github.com/figment-networks/indexing-engine v0.5.0
-	github.com/figment-networks/ni-cosmoslib/ibcmapper v0.0.6
 	github.com/gogo/protobuf v1.3.3
 	github.com/gravity-devs/liquidity v1.4.2
 )
