+++ api/mapper/freegrant.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	"github.com/figment-networks/indexing-engine/structs"
-	"github.com/figment-networks/ni-cosmoslib/figment/api/util"
+	"github.com/figment-networks/ni-cosmoslib/api/util"
 
 	"github.com/cosmos/cosmos-sdk/x/feegrant"
 	"github.com/gogo/protobuf/proto"
